+++ packages/rag-evaluation-site/src/components/organisms/index.ts
@@ -2,3 +2,5 @@ export * from './ContextDiagramPanel';
 export * from './TranscriptReaderPanel';
 export * from './AnnotationRailPanel';
 export * from './AnchoredCommentRail';
+export * from './CourseLessonPanel';
+export * from './CourseSlidePanel';